analytics_orchestrator/requirements.txt
# Analytics Orchestrator Dependencies

# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# HTTP Client and API
httpx==0.25.2
requests==2.31.0
aiohttp==3.9.1

# Database and Storage
sqlalchemy[asyncio]==2.0.23
aiosqlite==0.19.0
redis==5.0.1
sqlite3-utils==3.35

# Data Processing and Analytics
pandas==2.1.4
numpy==1.24.4
scipy==1.11.4
scikit-learn==1.3.2

# Configuration and Environment
pyyaml==6.0.1
python-dotenv==1.0.0
click==8.1.7

# Time and Date Handling
python-dateutil==2.8.2
pytz==2023.3

# Monitoring and Logging
prometheus-client==0.19.0
structlog==23.2.0

# Authentication and Security
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
authlib==1.2.1

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-httpx==0.27.0

# Development Tools
black==23.11.0
isort==5.12.0
mypy==1.7.1

# Web Framework Extras
jinja2==3.1.2
python-multipart==0.0.6